Move config data classes to separate folder
