+++ Extension/test/unitTests/includepath.test.ts
@@ -3,7 +3,7 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
 import * as Assert from "assert";
-import { IncludePath } from "../../src/extension/configprovider/includepath";
+import { IncludePath } from "../../src/extension/configprovider/data/includepath";
 
 suite("Test includepath parsers", () => {
     suite("Test compile output", () => {
